Removed relative time formatting due to inconsistencies.
While doing quality checks on usability and overall user experience, it turned out that the relative time formatting needs to be reworked in order to be understandable, usable and consistent. We're of course unhappy to remove a nice feature like this at this point, but it unavoidable.

Notify: gsd@otrs.com